Rename utils.py into _aux_task_caching.py
