+++ fractal/cmd/_task.py
@@ -12,8 +12,8 @@ from ..interface import BaseInterface
 from ..interface import PrintInterface
 from ..interface import RichJsonInterface
 from ..response import check_response
-from .utils import get_cached_task_by_name
-from .utils import refresh_task_cache
+from ._aux_task_caching import get_cached_task_by_name
+from ._aux_task_caching import refresh_task_cache
 
 
 async def get_task_list(client: AuthClient, **kwargs) -> RichJsonInterface:
